Provide SSH command to analyze H2 caches
Understanding how to configure cache-chroniclemap might be quite
difficult without having more information about the size of the caches
and average values of their keys and values.

Provide an SSH command so that cache-chroniclemap can be installed as a
plugin before being actually installed, in order to gather information
about the existing persisted H2 caches.

Bug: Issue 13475
Change-Id: I6f6ec28d68d9e5f9ff971a494bd45fd013e16d82