Short-circuit deserialization of invalid values
Values stored in persistent cache might become invalid over time (if the
expireAfterWrite value is enabled for example). The validation logic used
to be done by the chronicle-map cache implementation itself.

This had the disadvantage to of going through a complete deserialization
of the cached values just to invalidate them immediately after.

Short-circuit the deserialization phase by returning Invalid instances
in lieu of the original values, so that cache implementation can then
remove them from the cache.

This allows a faster processing of invalid values and, even more
crucially, sets the ground for validation logic that needs to happen
_before_ attempting deserialization (for example when version number
changed and the persisted value can no longer be deserialized with the
provided serializer).

Change-Id: Ic9983cc7df29724973b4ae75e6bac9b1b98c7ffc