Delegate to memory cache factory for non persistent caches
Restore the same contract compared to H2 based persistent caches: if
diskLimit is zero or negative, the cache is considered to be in memory
only and currently installed memory cache factory should be used.

According to this contract the factory for in memory caches can also be
overloaded using libModule.

Feature: Issue 13941
Change-Id: If911621de8ca9676d29850e449f14c684f73710e