Demonstrate overloading of backend for memory caches
Using CacheModule.MEMORY_MODULE module for memory caches can be
overloaded. It could also be done optionally, so that only the backend
for persistent caches is changed but not the backend for memory caches
or vice versa.

Test Plan:

Add this line to the <gerrit_site>/etc/gerrit.config file:

  installModule = com.googlesource.gerrit.modules.cache.chroniclemap.ChronicleMapInMemoryCacheModule

Now the backend for all in memory caches is Guava and not Caffeine.

Change-Id: I90031570badc9ad47b48ed1be4d891b08c7a9d58