Normalize chronicle-map configuration parameters
Chronicle-Map used to be instantiated not just for persistent caches,
but also for persistent caches that were effectively disabling
persistence by setting diskLimit to zero or a negative value.

This, however was fixed in[1] and since then the memory cache factory is
used for those caches.

The persistedFile is now passed in directly by the factory, so that
chronicle-map just cannot be instantiated without a persistent file to
store the data.

Consequently, some parameters that were required before to construct the
file path, such as name, version and diskLimit have now been removed.

[1]If911621de8ca9676d29850e449f14c684f73710e

Change-Id: I088be18d731371997f2af31ca3de8b80d36be18e