Remove diskLimit from configuration
Chronicle-Map used to be instantiated not just for persistent caches,
but also for persistent caches that were effectively disabling
persistence by setting diskLimit to zero or a negative value.

This, however was fixed in[1] and since then the memory cache factory is
used for those caches.

Since then, exposing 'diskLimit' via cache config is no longer required.

[1]If911621de8ca9676d29850e449f14c684f73710e

Change-Id: I088be18d731371997f2af31ca3de8b80d36be18e