Fix NPE thrown by LRU Cache
When attempting to remove a non-existing entry from the LRU cache, the
underlying synchronized map will return 'null' rather than a Boolean.

The attempt to unbox 'null' as a primitive boolean was causing an NPE
to be thrown.

Fix this by changing the return type of the remove() method to Boolean.

Bug: Issue 14099
Change-Id: If4aca686ed0fe1e4ccc2ad58deb50694a343dc38