Fix 'invalid metric name' error
MetricMaker only allows certain characters to be used when registering a
new metric name, raising an 'IllegalArgumentException' when the provided
metric contains invalid characters.

This in turns prevents the chronicle-map cache from being created.

Sanitize the name of the cache before registering the new metric, so
that potential invalid characters are replaced by '_'.

Bug: Issue 14098
Change-Id: Id62a0755245179c8d27f67a4a4c99b5f32bbeb72