Use existing pre-configured cache config for H2 migration
When migrating from H2 to ChronicleMap, it may be useful
to pre-define the target cache configuration beforehand.
That allows the Gerrit admin to correctly tune a desired
set of parameters and use them during the cutover from H2
to ChronicleMap.

Keep the default behaviour (auto-tune from H2) when the
average key and values sizes are not set, so to allow a
smooth migration for those Gerrit admins that do not
need to get the cutover fully automated.

Change-Id: I900cc0ed4a83fde6f77c13eb69c4212251834525