Skip unnecessary auto-adjustments of caches
Auto-adjusting the caches is an expensive operation
that could also occupy a considerable amount of extra
disk-space.

Re-building a cache that is already tuned would waste
precious resources and would not communicate effectively
whether a config change is needed or not.

Show only the config changes needed and rebuild only
the caches that need adjusting, leaving everything else
unchanged.

Bug: Issue 14876
Change-Id: Ic4d703d04410924c855ad2b96f8f635994d968a8