Front CacheChronicleMap with a Guava cache
The initial implementation of persistent caches using
ChronicleMap did not use any Guava cache front-end, relying
on the speed of accessing data on disk and memory.

Unfortunately, the serializer/deserializer of some entities
(e.g. change_notes) are very expensive in terms of CPU
and ChronicleMap does not store the binary objects as-is
but relies on in-memory persistence of keys and values.

Use the same structure of the H2 persistent cache where
the loader is delegated to the retrieval and deserialization
of the entities from ChronicleMap.

Change-Id: I8dd602c20fcd621972b562fc2163fefc7240484a