Front CacheChronicleMap with a Guava cache
The initial implementation of persistent caches using
ChronicleMap did not use any Guava cache front-end, relying
on the speed of accessing data on disk and memory.

Unfortunately, the deserialisation of some entities
(e.g. change_notes) is very expensive in terms of CPU
and ChronicleMap does not store the binary objects as-is
but relies on in-memory serialisation of keys and values.
There are two problems caused by this:
- delay in reading the cache from memory
- high system load when accessing the cache

Use the same structure of the H2 persistent cache where
the loader is delegated to the retrieval and deserialisation
of the entities from ChronicleMap.

Amend the cache statistics and return the in-memory elements
and the ones stored on disk as separate values.

Bug: Issue 15036
Change-Id: I8dd602c20fcd621972b562fc2163fefc7240484a