Avoid full cache scanning for pruning
The full cache scanning should be avoided at all costs
because of the massive CPU and JVM heap utilisation due
to the deserialisation of the cached values.

The prune() of the cache cannot rely on any ordering
of entries and therefore needs another data structure
to remember the historical insertion ordering of the
keys.

Use an in-memory queue of keys and use it for scanning
them in historical order for expirying them for maxAge.

TODO: The queue is currently in memory which isn't going
to work across restarts. The queue needs to be persisted
to disk as a circular file with automatic reorganiation.

Bug: Issue 15121
Change-Id: Ifa3068e2d27713b452c4bcab6a611e3c9b511868