Allow to specify number of entries in auto-adjust-caches
When running the auto-adjust-caches command, either via SSH or via HTTP,
it is not possible to specify the maximum number of entries for the
resulting cache.

This means, that the maximum number of entries allowed in the cache
cannot be accommodated to make room for caches that are increasing in
size.

Introduce the `max-entries` parameter to tell the auto-adjust-caches
the maximum number of entries desired in the resulting auto-tuned cache.

Bug: Issue 15412
Change-Id: I0ff43060a8fd2e22c3a60d11dcec1f875ba45e6e