[DO NOT SUBMIT] Allow to specify number of entries in auto-adjust-caches
When running the auto-adjust-caches command, either via SSH or via HTTP,
it is not possible to specify the maximum number of entries for the
resulting cache.

This means, that the maximum number of entries allowed in the cache
cannot be accommodated to make room for caches that are increasing in
size.

Introduce the `max-entries` parameter to tell the auto-adjust-caches
the maximum number of entries desired in the resulting auto-tuned cache.

Fix link to auto-adjust caches in tuning docs

NOTE: This is supposed to just show how the fixups on Change 326200
      look like.

Original-Author: syntonyze@gmail.com
Change-Id: I36325faf743f3aa294eff9d48f2bda1559f5c717