Expose max auto resizes metric
urrently, only the number of auto resizes is exposed as a metric.
When this value reaches 0 it is no longer possible to
expand the cache.

However, it is not possible to know the maximum number of auto-resizes
that the cache could go through.

Expose the maximum number of auto-resizes as a metric, so that can be
used together with the remaining auto-resizes to evaluate the overall
utilization of the cache.

Bug: Issue 15458
Change-Id: Ie81684950a28f3e2420bddfcaa10b7db218d00fb