Do not throw when failing to store into chronicle-map
When storing a new entry in chronicle-map it is possible that the
operation might fail.

For instance, if the new entry diverges greatly from the configured
average key and value size, the available memory chunks might not be
enough to accommodate the new entry.

Chronicle-map fails with:

```
Entry is too large: requires X chunks, Y is the maximum.
```

Another typical case would be that the cache already reached the maximum
number of auto-resizes and thus it is not able to expand anymore.

Chronicle-map fails with:

```
Attempt to allocate #X extra segment tier, Y is maximum.
```

In these cases, as well as other possible failures, we don't want the
exception to bubble up to the stack and cause the caller to fail.

The caller might be a REST API initiated by the UI via ajax, which in
turn, would cause the UI to stop rendering and a bad user experience.

Catch chronicle-map's `put()` exception locally and simply avoid caching
the offending entry.

Bug: Issue 15483
Change-Id: Ia1811d66b676388f15cb6f42ef3ee4fd5d1aa94b