Move store related metrics to ChronicleMapStore
Do not emit all metrics from the chronicle-map cache implementation, but
rather split them so that the cache only emits cache-related metrics,
whilst the store emits storage-related ones.

This unlocks the possibility to centralize the emission of
store-related metrics regardless whether the store is accessed from
the ChronicleMapCacheImpl or the ChronicleMapCacheLoader.

Change-Id: Ie5e50a79f48f6e1e1ae4845b899bd9ad52e4c638