Postpone call to get() when loading from the in-memory cache
Caffeine's get() method locks on read.
This is unexpected and possibly a bug with the caffeine implementation.

The effect of this on cache-chroniclemap is that when multiple threads
try to access the in-memory cache concurrently, they result in being
BLOCKED.

This, in turn, causes high consumption of Gerrit thread, which might
eventually, make it unresponsive.

Workaround this problem by first trying to get the key _if present_ and
only calling get() in case no value could be loaded.

This works because getIfPresent(), as opposed to, get(), does not lock
on read and thus decreases the chances of thread blocking.

Bug: Issue 15645
Change-Id: I7021d93d4011e252c5c022ffcd3eecee22820ff5