Record the total number of failures to put entries across all stores
Expose a counter called 'cache/chroniclemap/store_put_failures'
for tracking the total number of failures in putting entries across
all cache-chroniclemap stores for all caches.

The rationale of the new cumulative counter is allowing a simple
monitoring and alerting when any of the store is going into trouble
because of failures to add entries to it.

Change-Id: I247422bdb64e47706176ae0bbd503f9e075aba5e