Avoid full cache scanning for pruning
The full cache scanning should be avoided at all costs
because of the massive CPU and JVM heap utilisation due
to the deserialisation of the cached values.

The prune() of the cache cannot rely on any ordering
of entries and therefore needs another data structure
to remember the historical insertion ordering of the
keys.

Use an in-memory linked hash set so that MRU keys are propagated by
remove-add on each key access to the tail of it whereas LRU key is
always available as a head key (therefore it is cheap to remove LRU
keys). In addition to key store also the value insertion time so that
expiring entries for maxAge is also possible.

TODO: The queue is currently in memory which isn't going
to work across restarts. The queue needs to be persisted
to disk as a circular file with automatic reorganiation.

Bug: Issue 15121
Change-Id: Ifa3068e2d27713b452c4bcab6a611e3c9b511868