TimeValueMarshaller: use higher level API for (de)serialization
At presnt each read/write call results in buffers being created at
least once for write but more for read. Use higher level API in order to
avoid buffer creation and leave it up to the implementation to optimize
on it.

Note that TimedValue was turned to mutable object to leverage on
re-suing objects and furthere reduce the memory footprint (see [1])

[1] Read interface documentation excerpt:
(...) should attempt to reuse the given using object, i. e. to read
the deserialized data into the given object. If it is possible, this
object then returned from this method back. If it is impossible for
any reason, a new object should be created and returned (...)

Change-Id: Ifb194e49b9bace0cf4d81ef3af3d6a64f0564dbd