Remove metrics when cache gets closed
Issue:
When plugin with consistent caches gets reloaded it fails to load with
the following exception:

  java.lang.IllegalArgumentException: A metric named cache/chroniclemap/percentage_free_space_test-cache-plugin_test_cache already exists
  ...

The problem is caused by metrics that are not removed when previous
version of plugin gets unloaded.

Solution:
Make sure that all cache related metrics are removed when cache is
closed.

Note that this is the necessary but not sufficient step to solve #15587.
It is enough to solve the happy path when 'Gerrit-ReloadMode: restart'.
IOW when old version of plugin gets unloaded before new version is
loaded.

Bug: Issue 15587
Depends-On: I06d48535702a8acd0bcb4e1315a35b3c1d1768cb
Change-Id: I03e34dd3463fad00ba8fbc174597c896a93db547