Introduce metric for caches that fall back to default config
Cache chroniclemap relies heavily on configuration (especially on
average key and values sizes) so that its performance can be optimal.
What is more starting with `stable-3.4` there are 7 core caches that
are missing their confifuration in Defaults and use fall-back
configuration. It is strongly adviced to auto-tune them (see
'tuning.md') and store dedicated configuration in 'gerrit.config'.

Issuing a warning in such case is not a perfect solution as Gerrit core
caches are initiated very early during the server start hence messages
are not always visible in the log. For this reason the following metric
was introduced:

* 'cache/chroniclemap/caches_without_configuration' that could be used
  as an alarm source as it contains the total number of caches that
  have no dedicated configuration in 'gerrit.config' and use defaults or
  fall-back configuration

Notes:
* adding cache configuration to `gerrit.config` is not considered as
  default configuration
* additionally, the 'auto-adjust-caches' command was extended with
  'adjust-caches-on-defaults' option (see the tuning.md for details)
  that it can be used to tune all caches that fall back to defaults

Bug: Issue 15865
Change-Id: I8b0835c6d8a8ae96832371b79afd2a11f69f09b9