Introduce metrics for caches that falls back to default config
Cache chroniclemap relies heavily on configuration (especially on
average key and values sizes) so that its performance can be optimal.
However starting with `stable-3.4` there are 7 caches that uses
fall-back (default) configuration that is/might be sub-optimal from the
severl operation point of view.

Issueing a warning in such case is not a perfect solution as Gerrit core
caches are initiated very early during the server start hence messages
are not always visible in the log. For this reason the following metrics
were introduced:

* 'cache/chroniclemap/caches_without_chroniclemap_configuration' that
  could be used as an alarm source as it contians the total number of
  caches that fall back to default configuation

* 'cache/chroniclemap/cache_without_chroniclemap_configuration_<cache-name>'
  that is created only for cache that falls back to defaults so that
  misconfigured caches could be easily identified when the previous
  metric is > 0

Bug: Issue 15865
Change-Id: I8b0835c6d8a8ae96832371b79afd2a11f69f09b9