Fix eviction logic during prune phase
When the prune is triggered, the eviction logic was not stopping
at the expected threashold but cleared out most of the cache entries.
The breakage was not detected because of incomplete test coverage
or invalid test conditions.

Fix the eviction tests and the eviction logic, so that the cache
remains full as much as possible.

Change-Id: Ib3546c46e8cced40a9f8b08aeac23961f108c353