Ensure JDK 17 support.
The cache-chroniclemap internals are JDK 17 compatible however JDK,
since that version is more restrictive on which modules are by default
accessible to the third party libraries. Enhance plugin's README.md
and build.md with steps that are necessary to run module in Gerrit
under JDK 17 JVM and call test from IDE (as described in [1]).

In addition JDK 17 modifies the way how annotations have their
`toString` method implemented. In JDK 11 (and older) it used to be:

  @com.google.inject.name.Named(value="accounts")

Whereas in JDK 17 it is:

  @com.google.inject.name.Named("accounts")

Modify MigrateH2CachesLocalDiskIT.isClassBoundWithName(Entry) to be
compatible with both implementations.

[1] https://chronicle.software/chronicle-support-java-17/

Bug: Issue 306464780
Change-Id: I3e201d9638de7007aca2af10796a8ee1deb9d04c