Fix compilation error in SSHCommandModule
After change I35f1a785, the base class of SSHComandModule, the
PluginComandModule API has changed and now takes a plugin name as a
constructor argument. To conform to that change we inject the plugin
name and pass it to the parent constructor.

Also removes duplicated static import of `assertThat` in
ChronicleMapCacheIT.

Bug: Issue 331383932
Change-Id: I4c12efa489665246932c6c65ab9e830c9dc18478