Fix test breakages on Gerrit master
The latest Gerrit master does not preserve anymore
the SSH sessions in the test context and instead is sharing
them across tests.

The new behaviour prevents the execution of parallel
integration tests, forcing the cache-chroniclemap tests
to use a self-contained mechanism for allocatign and releasing
sessions without having to use the one predefined in Gerrit
test infrastructure.

Introduce a new LightweightPluginDaemonWithSshSessionsTest
base class for all tests in cache-chroniclemap that need
to allocate a SSH session to talk to Gerrit.

Change-Id: I071cdcec5b1224974ce638fad58cba244fc31e2c