Enable cache-chroniclemap to Run on Java 21
With Chronicle 25, OpenHFT has introduced official support for Java 21,
which we’re adopting here to allow cache-chroniclemap to run an a Gerrit
3.11 on Java 21[1].

Since Chronicle 24, OpenHFT has limited Maven Central to early access
(ea) releases, with stable releases now exclusive to customers with a
support agreement [2]. As a result, 3.23.5 is the latest stable release
publicly available on Maven Central [3].

Per OpenHFT guidance [2]:

``` Many organizations use the ‘ea’ release in production, as this
contains the latest functionality and fixes.  ```

Following this recommendation, we’re updating to the latest ea release
in the Chronicle 25 series, which is the earliest version to introduce
Java 21 support.

[1] https://chronicle.software/chronicle-25-whats-new-and-improved/
[2] https://github.com/OpenHFT/OpenHFT/blob/aa79b18031dbecfd6722aa285a495e87e2d016d5/docs/Version-Support.adoc#ea-early-access
[3] https://mvnrepository.com/artifact/net.openhft/chronicle-map

Bug: Issue 377724725
Change-Id: I40bcf7994798297ef371a1562ea4cd4df055901d