Set the DiskStats.invalidatedCount to zero
The tracking of the invalidation count was introduced
in Ie82d95df8aa for enabling the automatic rebuild of the
BloomFilter is used in the H2 persistent cache.

Cache-Chroniclemap does not expand or shrink its cache and
simply keep an index of the items that are valid or invalid
without having to rebuild it.

The additional invalidation count is therefore not useful
in this implementation, it can be safely set to zero.

Change-Id: Ic1b35d1307a8a470240c6f4d75114831015c4e1e